+++ src/main/java/com/capstone/simplek/Model/Service.java
@@ -1,6 +1,4 @@
-package com.capstone.simplek.Service;
-import com.capstone.simplek.Children.Children;
-import com.capstone.simplek.School.School;
+package com.capstone.simplek.Model;
 import javax.persistence.*;
 import java.util.List;
