+++ src/main/java/com/capstone/simplek/Repository/UserRepository.java
@@ -1,4 +1,5 @@
-package com.capstone.simplek.User;
+package com.capstone.simplek.Repository;
+import com.capstone.simplek.Model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
